Repos/Backend/pyproject.toml
[project]
name = "wgucapstonebackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite==0.20.0",
    "annotated-types==0.7.0",
    "anyio==4.6.0",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asttokens==2.4.1",
    "async-lru==2.0.4",
    "asyncpg==0.30.0",
    "attrs==24.2.0",
    "babel==2.16.0",
    "bcrypt==4.2.0",
    "beautifulsoup4==4.12.3",
    "bleach==6.1.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "colorama==0.4.6",
    "comm==0.2.2",
    "contourpy==1.3.0",
    "cycler==0.12.1",
    "debugpy==1.8.6",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "executing==2.1.0",
    "fastapi[standard]==0.115.0",
    "fastapi-cli==0.0.5",
    "fastjsonschema==2.20.0",
    "fonttools==4.54.1",
    "fqdn==1.5.1",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httpx==0.27.2",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==8.28.0",
    "isoduration==20.11.0",
    "jedi==0.19.1",
    "jinja2==3.1.4",
    "jiter==0.6.1",
    "joblib==1.4.2",
    "json5==0.9.25",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2023.12.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyter-events==0.10.0",
    "jupyter-lsp==2.2.5",
    "jupyter-server==2.14.2",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.2.5",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "kiwisolver==1.4.7",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "matplotlib==3.9.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.0.2",
    "nbclient==0.10.0",
    "nbconvert==7.16.4",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "notebook==7.2.2",
    "notebook-shim==0.2.4",
    "numpy==2.1.2",
    "openai==1.52.2",
    "overrides==7.7.0",
    "packaging==24.1",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pillow==10.4.0",
    "platformdirs==4.3.6",
    "prometheus-client==0.21.0",
    "prompt-toolkit==3.0.48",
    "psutil==6.0.0",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pydantic-settings==2.6.1",
    "pygments==2.18.0",
    "pyjwt==2.9.0",
    "pyparsing==3.1.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-json-logger==2.0.7",
    "python-multipart==0.0.12",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "referencing==0.35.1",
    "requests==2.32.3",
    "resend==2.4.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rich==13.9.2",
    "rpds-py==0.20.0",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "seaborn==0.13.2",
    "send2trash==1.8.3",
    "setuptools==75.6.0",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.35",
    "sqlmodel==0.0.22",
    "stack-data==0.6.3",
    "starlette==0.38.6",
    "terminado==0.18.1",
    "threadpoolctl==3.5.0",
    "tinycss2==1.3.0",
    "tornado==6.4.1",
    "tqdm==4.66.5",
    "traitlets==5.14.3",
    "typer==0.12.5",
    "types-python-dateutil==2.9.0.20241003",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "ucimlrepo==0.0.7",
    "uri-template==1.3.0",
    "urllib3==2.2.3",
    "uvicorn==0.31.0",
    "watchfiles==0.24.0",
    "wcwidth==0.2.13",
    "webcolors==24.8.0",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "websockets==13.1",
    "wordcloud==1.9.3",
]
